lib/http: split public include headers
