+++ lib/http/tcp_client.c
@@ -1,6 +1,6 @@
-#include "tcp.h"
-#include "tcp_internal.h"
+#include "http/tcp.h"
 #include "sock.h"
+#include "tcp_internal.h"
 
 #include "logging.h"
