+++ lib/http/url.c
@@ -1,5 +1,6 @@
-#include "url.h"
 #include "parse.h"
+
+#include "http/url.h"
 
 #include "logging.h"
